container-with-most-water/src/test/java/com/github/tulesaza/containerwithmostwater/SolutionTest.java
package com.github.tulesaza.containerwithmostwater;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

class SolutionTest {

    final Solution solution = new Solution();

    @ParameterizedTest
    @MethodSource("data")
    void maxArea(int[] heights, int expected) {
        assertEquals(expected, solution.maxArea(heights));
    }

    static Stream<Arguments> data(){
        return Stream.of(
                Arguments.of(new int[]{1,8,6,2,5,4,8,3,7}, 49),
                Arguments.of(new int[]{1,1}, 1)
        );
    }
}